Relationships

+++ database/migrations/2019_12_25_195738_create_replies_table.php
@@ -16,7 +16,7 @@ class CreateRepliesTable extends Migration
         Schema::create('replies', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->text('body');
-            $table->integer('question_id')->unsigned();
+            $table->unsignedBigInteger('question_id')->unsigned();
             $table->integer('user_id')->unsigned();
             $table->foreign('question_id')->references('id')->on('questions')
             ->onDelete('cascade');
